add validation parameter ID

+++ models/employeeModel.js
@@ -58,21 +58,26 @@ const getPartner = (request, response) => {
   result.success = false
   result.message = ''
   result.data = []
-  pool.query('SELECT e_req.employee_id, e_req.employee_email, e_req.employee_id_no, e_req.employee_name, e_req.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_app.employee_id_no = $1 UNION ALL SELECT e_app.employee_id, e_app.employee_email, e_app.employee_id_no, e_app.employee_name, e_app.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_req.employee_id_no = $1',[id_no],(error, results) => {
-    if (error) {
-      result.message = 'Failed get data employee, please try again'
-      response.status(500).json(result)
-    }
-    result.success = true
-    result.message = results.rowCount+' data found'
-    if(results.rowCount > 0){
-      result.data = results.rows
-    } else {
-      result.message = 'Data not found'
-      result.data = []
-    }
-    response.status(200).json(result)
-  })
+  if(Number.isInteger(id_no) === false){
+    result.message = `Invalid parameter ID No `
+    return response.status(400).json(result);
+  } else {
+    pool.query('SELECT e_req.employee_id, e_req.employee_email, e_req.employee_id_no, e_req.employee_name, e_req.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_app.employee_id_no = $1 UNION ALL SELECT e_app.employee_id, e_app.employee_email, e_app.employee_id_no, e_app.employee_name, e_app.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_req.employee_id_no = $1',[id_no],(error, results) => {
+      if (error) {
+        result.message = 'Failed get data employee, please try again'
+        response.status(500).json(result)
+      }
+      result.success = true
+      result.message = results.rowCount+' data found'
+      if(results.rowCount > 0){
+        result.data = results.rows
+      } else {
+        result.message = 'Data not found'
+        result.data = []
+      }
+      response.status(200).json(result)
+    })
+  }
 }
 
 const getPartnerOfPartner = (request, response) => {
@@ -80,21 +85,26 @@ const getPartnerOfPartner = (request, response) => {
   result.success = false
   result.message = ''
   result.data = []
-  pool.query('SELECT e_req.employee_id, e_req.employee_email, e_req.employee_id_no, e_req.employee_name, e_req.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_app.employee_id_no = $1 UNION ALL SELECT e_app.employee_id, e_app.employee_email, e_app.employee_id_no, e_app.employee_name, e_app.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_req.employee_id_no = $1',[id_no],(error, results) => {
-    if (error) {
-      result.message = 'Failed get data employee, please try again'
-      response.status(500).json(result)
-    }
-    result.success = true
-    result.message = results.rowCount+' data found'
-    if(results.rowCount > 0){
-      result.data = results.rows
-    } else {
-      result.message = 'Data not found'
-      result.data = []
-    }
-    response.status(200).json(result)
-  })
+  if(Number.isInteger(id_no) === false){
+    result.message = `Invalid parameter ID No `
+    return response.status(400).json(result);
+  } else {
+    pool.query('SELECT e_req.employee_id, e_req.employee_email, e_req.employee_id_no, e_req.employee_name, e_req.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_app.employee_id_no = $1 UNION ALL SELECT e_app.employee_id, e_app.employee_email, e_app.employee_id_no, e_app.employee_name, e_app.employee_phone FROM employee_partner ep INNER JOIN employee e_req ON ep.employee_id_request = e_req.employee_id INNER JOIN employee e_app ON ep.employee_id_approved = e_app.employee_id WHERE e_req.employee_id_no = $1',[id_no],(error, results) => {
+      if (error) {
+        result.message = 'Failed get data employee, please try again'
+        response.status(500).json(result)
+      }
+      result.success = true
+      result.message = results.rowCount+' data found'
+      if(results.rowCount > 0){
+        result.data = results.rows
+      } else {
+        result.message = 'Data not found'
+        result.data = []
+      }
+      response.status(200).json(result)
+    })
+  }
 }
 
 const getEmployeePartnerCompany = (request, response) => {
@@ -103,86 +113,86 @@ const getEmployeePartnerCompany = (request, response) => {
   result.message = ''
   result.data = {}
 
-  // let cb = (resx) => {
-  //   console.log(resx);
-  //   return resx;
-  // }
-  // let x = pool.query('select * from company')
-  //   .then( res=> {
-  //     return cb(res)
-  //   })
-  //   .catch( err =>{
-  //     return err
-  //   })
-
-
-  // return x.then( y => { return y} );
-
-  pool.query('SELECT c.company_id, c.company_code, c.company_name, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce, employee e, company c WHERE ce.employee_id = e.employee_id AND ce.company_id = c.company_id AND e.employee_id_no = $1',[id_no],(error, getCompany) => {
-    if (error) {
-      result.message = 'Failed get data company, please try again'
-      response.status(500).json(result)
-    }
-
-    if(getCompany.rowCount > 0) {
-
-      getCompany.rows.forEach(element => {
-        element.partner_employee = []
-        var company_id = element.company_id
-        var join_date = element.join_date
-
-        pool.query('SELECT e.employee_id, e.employee_email, e.employee_name, e.employee_id_no, e.employee_phone, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce, employee e WHERE ce.employee_id = e.employee_id AND ce.company_id=$2 AND e.employee_id_no <> $1 AND ce.join_date >= $3',[id_no,company_id,join_date],(error, getPartnerEmployeeCompany) => {
-          if (error) {
-            result.message = 'Failed get data employee, please try again'
-            response.status(500).json(result)
-          }
-          try{
-            element.partner_employee = getPartnerEmployeeCompany.rows
-            console.log(element)
-          }catch(e){
-            console.log(e)
-          }
+  if(Number.isInteger(id_no) === false){
+    result.message = `Invalid parameter ID No `
+    return response.status(400).json(result);
+  } else {
+    pool.query('SELECT c.company_id, c.company_code, c.company_name, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce, employee e, company c WHERE ce.employee_id = e.employee_id AND ce.company_id = c.company_id AND e.employee_id_no = $1',[id_no],(error, getCompany) => {
+      if (error) {
+        result.message = 'Failed get data company, please try again'
+        response.status(500).json(result)
+      }
+  
+      if(getCompany.rowCount > 0) {
+  
+        getCompany.rows.forEach(element => {
+          element.partner_employee = []
+          var company_id = element.company_id
+          var join_date = element.join_date
+          
+          return new Promise(function(resolve,reject){
+            pool.query('SELECT e.employee_id, e.employee_email, e.employee_name, e.employee_id_no, e.employee_phone, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce, employee e WHERE ce.employee_id = e.employee_id AND ce.company_id=$2 AND e.employee_id_no <> $1 AND ce.join_date >= $3',[id_no,company_id,join_date],(error, getPartnerEmployeeCompany) => {
+              if (error) {
+                result.message = 'Failed get data employee, please try again'
+                response.status(500).json(result)
+              }
+              try{
+                element.partner_employee = getPartnerEmployeeCompany.rows
+                resolve(element)
+                resolve(getCompany.rows)
+                console.log(element)
+              }catch(e){
+                console.log(e)
+                reject(e)
+              }
+            })
+          })
         })
-      })
-    } 
-
-    result.success = true
-    result.message = getCompany.rowCount+' data found'
-    if(getCompany.rowCount > 0){
-      result.data = getCompany.rows
-    } else {
-      result.message = 'Data not found'
-      result.data = []
-    }
-    response.status(200).json(result)
-  })
+      } 
+  
+      result.success = true
+      result.message = getCompany.rowCount+' data found'
+      if(getCompany.rowCount > 0){
+        result.data = getCompany.rows
+      } else {
+        result.message = 'Data not found'
+        result.data = []
+      }
+      response.status(200).json(result)
+    })
+  }
 }
 
 const getHistoryCompanyEmployee = (request, response) => {
   const id_no = parseInt(request.params.id_no)
   result.success = false
   result.data = {}
-  pool.query('SELECT c.company_id,c.company_name, c.company_address, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce INNER JOIN employee e ON ce.employee_id = e.employee_id INNER JOIN company c ON ce.company_id = c.company_id WHERE e.employee_id_no = $1 AND c.is_deleted = false', [id_no], (error, getCompany) => {
-    if (error) {
-        result.message = 'Invalid parameter ID'
-        response.status(500).json(result)
-    } else {
-      result.success = true
-      result.data.previous_company = []
-      result.data.current_company = []
-      if(getCompany.rowCount > 0) {
-        getCompany.rows.forEach(element => {
-          if(element.resign_date){
-            result.data.previous_company.push(element)
-          } else {
-            result.data.current_company.push(element)
-          }
-        });
-        result.message = getCompany.rowCount+' data found'
+  if(Number.isInteger(id_no) === false){
+    result.message = `Invalid parameter ID No `
+    return response.status(400).json(result);
+  } else {
+    pool.query('SELECT c.company_id,c.company_name, c.company_address, ce.division, ce.job_position, ce.join_date, ce.resign_date FROM company_employee ce INNER JOIN employee e ON ce.employee_id = e.employee_id INNER JOIN company c ON ce.company_id = c.company_id WHERE e.employee_id_no = $1 AND c.is_deleted = false', [id_no], (error, getCompany) => {
+      if (error) {
+          result.message = 'Invalid parameter ID'
+          response.status(500).json(result)
+      } else {
+        result.success = true
+        result.data.previous_company = []
+        result.data.current_company = []
+        if(getCompany.rowCount > 0) {
+          getCompany.rows.forEach(element => {
+            if(element.resign_date){
+              result.data.previous_company.push(element)
+            } else {
+              result.data.current_company.push(element)
+            }
+          });
+          result.message = getCompany.rowCount+' data found'
+        }
+        response.status(200).json(result)
       }
-      response.status(200).json(result)
-    }
-  })
+    })
+  }
 }
 
 const searchEmployeeByName = (request, response) => {
@@ -273,8 +283,11 @@ const updateEmployee = (request, response) => {
     result.success = false
     result.message = ''
     result.data = []
-
-    pool.query('SELECT employee_id FROM employee WHERE (employee_id_no = $2 OR employee_email = $3) AND is_deleted = false AND employee_id != $1', [id, employee_id_no, employee_email], (error, checkEmailId) => {
+    if(Number.isInteger(id) === false){
+      result.message = `Invalid parameter ID No `
+      return response.status(400).json(result);
+    } else {
+      pool.query('SELECT employee_id FROM employee WHERE (employee_id_no = $2 OR employee_email = $3) AND is_deleted = false AND employee_id != $1', [id, employee_id_no, employee_email], (error, checkEmailId) => {
         
         if(checkEmailId.rowCount > 0) {
             result.message = `ID No : ${employee_id_no}  or Email : ${employee_email}  is already exist `
@@ -295,22 +308,28 @@ const updateEmployee = (request, response) => {
             )
         }
     })
+    }
 }
 
 const deleteEmployee = (request, response) => {
     const id = parseInt(request.params.id)
-    result.success = false
-    result.message = ''
-    result.data = []
-    pool.query('DELETE FROM employee WHERE employee_id = $1', [id], (error, results) => {
-        if (error) {
-            result.message = 'Invalid parameter ID'
-            response.status(500).json(result)
-        }
-        result.success = true
-        result.message = `Employee deleted with Company ID: ${id}`
-        return response.status(200).json(result)
-    })
+    if(Number.isInteger(id) === false){
+      result.message = `Invalid parameter ID No `
+      return response.status(400).json(result);
+    } else {
+      result.success = false
+      result.message = ''
+      result.data = []
+      pool.query('DELETE FROM employee WHERE employee_id = $1', [id], (error, results) => {
+          if (error) {
+              result.message = 'Invalid parameter ID'
+              response.status(500).json(result)
+          }
+          result.success = true
+          result.message = `Employee deleted with Company ID: ${id}`
+          return response.status(200).json(result)
+      })
+    }
 }
 
 module.exports = {
